python/otel_prometheus_kafka/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
confluent-kafka==2.5.0
Deprecated==1.2.14
googleapis-common-protos==1.63.2
grpcio==1.66.0
idna==3.8
importlib_metadata==8.0.0
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-exporter-prometheus==0.47b0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-confluent-kafka==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
prometheus_client==0.20.0
protobuf==4.25.4
Pympler==1.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
wrapt==1.16.0
zipp==3.20.0
